fix (be): fix logic payment in create payment and pay remainder

+++ BE/src/main/java/com/project/evrental/domain/dto/response/SearchMetadata.java
@@ -14,7 +14,7 @@ public class SearchMetadata {
 
     Integer totalResults;
     
-    Integer radiusKm;
+    Double radiusKm;
     
     Integer returnedCount;
